print debug text for dwithin and arccos errors
